test1

+++ src/busybox/Mv.java
@@ -3,11 +3,16 @@ package busybox;
 import java.io.File;
 
 public class Mv {
-	 public void mv(String oldpath,String newpath){
-	     if(!oldpath.equals(newpath)){
-	         File oldfile=new File(oldpath);
-	         File newfile=new File(newpath);
-	         oldfile.renameTo(newfile);
-	         }
+	 public static void mv(String oldpath,String newpath){
+	 	if(!oldpath.equals(newpath)){
+	 		File oldfile=new File(oldpath);
+	 		if (oldfile.exists()||oldfile.isDirectory()){
+				File newfile=new File(newpath);
+				oldfile.renameTo(newfile);
+				System.out.println("移动成功");
+			}else{
+				System.out.println("未找到该文件");
+			}
+	 	}
 	 }
 }
